Make master default branch name.
This is not because I'm a racist pig, it's just that muscle memory
writes master, so as long as it's the mainstream thing in other projects
i will be using master. It's a more useful analogy in distributed
systems than git, but I don't think this naming is important and
therefore I won't make a thing of it.